requirements.txt
Flask
inference_sdk
roboflow
opencv-python
numpy
Pillow
requests
gunicorn
